adds identity details
