+++ workspaces/idapp/app/(tabs)/address-book/identity-details.tsx
@@ -1,0 +1,5 @@
+import { IdentityDetails } from '~/views/address-book'
+
+export default function Identities() {
+  return <IdentityDetails />
+}
